+++ careerBridge/src/main/java/com/careerBridge/careerBridge/service/InternshipService.java
@@ -1,30 +1,50 @@
 package com.careerBridge.careerBridge.service;
 
-import com.careerBridge.careerBridge.entity.Company;
+
 import com.careerBridge.careerBridge.entity.Internship;
+import com.careerBridge.careerBridge.repository.CompanyRepository;
 import com.careerBridge.careerBridge.repository.InternshipRepository;
 import com.careerBridge.careerBridge.dto.InternshipRequest;
 import com.careerBridge.careerBridge.exception.ResourceNotFoundException;
-import org.springframework.beans.factory.annotation.Autowired;
+import com.careerBridge.careerBridge.security.SecurityService;
 import org.springframework.stereotype.Service;
+import com.careerBridge.careerBridge.entity.Company;
+import org.springframework.security.access.AccessDeniedException;
+import com.careerBridge.careerBridge.entity.User;
+import com.careerBridge.careerBridge.entity.Role;
 import java.util.List;
 import java.time.LocalDateTime;
 
 @Service
 public class InternshipService {
-    @Autowired
-    private InternshipRepository internshipRepository;
+
+    private final InternshipRepository internshipRepository;
+    private final CompanyRepository companyRepository;
+    private final SecurityService securityService;
+
+    public InternshipService(InternshipRepository internshipRepository, CompanyRepository companyRepository, SecurityService securityService) {
+        this.internshipRepository = internshipRepository;
+        this.companyRepository = companyRepository;
+        this.securityService = securityService;
+    }
 
     public Internship saveInternship(InternshipRequest request) {
 
+        User currentUser = securityService.getCurrentUser();
+
+        Company company = companyRepository.findByUserId(currentUser.getId())
+                .orElseThrow(() ->
+                        new ResourceNotFoundException("Company profile not found"));
+
         Internship internship = new Internship();
+
+        internship.setCompany(company);
 
         internship.setCreatedAt(LocalDateTime.now());
         internship.setTitle(request.getTitle());
         internship.setDescription(request.getDescription());
         internship.setRequirements(request.getRequirements());
         internship.setLocation(request.getLocation());
-        internship.setCompanyName(request.getCompanyName());
         internship.setType(request.getType());
         internship.setDurationInMonths(request.getDurationInMonths());
         internship.setStipend(request.getStipend());
@@ -47,11 +67,12 @@ public class InternshipService {
     public Internship updateInternship(Long id, Internship internship) {
         Internship existing = getInternshipById(id);
 
+        verifyOwnership(existing);
+
         existing.setTitle(internship.getTitle());
         existing.setDescription(internship.getDescription());
         existing.setRequirements(internship.getRequirements());
         existing.setLocation(internship.getLocation());
-        existing.setCompanyName(internship.getCompanyName());
         existing.setType(internship.getType());
         existing.setDurationInMonths(internship.getDurationInMonths());
         existing.setStipend(internship.getStipend());
@@ -65,6 +86,21 @@ public class InternshipService {
     public void deleteInternshipById(Long id) {
         Internship internship=internshipRepository.findById(id).orElseThrow(()->
                 new ResourceNotFoundException("Internship not found"));
+        verifyOwnership(internship);
         internshipRepository.delete(internship);
     }
+
+    private void verifyOwnership(Internship internship) {
+        User currentUser = securityService.getCurrentUser();
+
+        if(currentUser.getRole()==Role.ADMIN){
+            return;
+        }
+
+        if(!currentUser.getId().equals(internship.getCompany().getUser().getId())) {
+            throw new AccessDeniedException("You do not have permission to modify this Internship");
+        }
+
+    }
+
 }
